==> 9_snippets.ipynb <==
import logging
logging.basicConfig(format='%(asctime)s - %(levelname)s - %(message)s', level=logging.INFO)
MIN = 10
MAX = 20

def is_ok(pauses, duration, MIN=10, MAX=20):

    pauses = [0, *pauses, duration]

    for start, end in zip(pauses[:-1], pauses[1:]):
        if not MIN <= end - start <= MAX:
            return False
    return True


def splitter(pauses, duration, MIN=10, MAX=20):
    from itertools import combinations, chain

    def powerset(iterable):
        "powerset([1,2,3]) --> () (1,) (2,) (3,) (1,2) (1,3) (2,3) (1,2,3)"
        s = list(iterable)
        return chain.from_iterable(combinations(s, r) for r in range(len(s)+1))
    all_combinations = list(powerset(pauses))
    logging.info(f"Testing {len(all_combinations)} combinations....")
    for combination in all_combinations:
        if is_ok(combination, duration, MIN=MIN, MAX=MAX):
            return list(combination)
    logging.warning(f"No solution found so that {MIN=}s <= duration <= {MAX=}s.")
    return None


pauses = [7, 15, 16, 20]
splitter(pauses, 28, MIN=5, MAX=10)


def splitter2(pauses, duration, MIN=10, MAX=20):
    pauses = [0, *pauses, duration]
    subsequent_pauses = [[s,e] for s, e in zip(pauses, pauses[1:]) if MIN <= e-s <= MAX]
    res = set()
    for s, e in subsequent_pauses:
        if s != 0:
            res.add(s)
        if e != duration:
            res.add(e)
    return sorted(list(res))
    


pauses = [7, 15, 16, 20]
splitter2(pauses, 28, MIN=5, MAX=10)

from pydub import AudioSegment
from pydub.silence import split_on_silence, detect_silence
file = "/home/peterr/macocu/task6_speech/data/00009875.flac.wav"
sound_file = AudioSegment.from_wav(file)
audio_chunks = split_on_silence(sound_file, min_silence_len=1000, silence_thresh=-40 )
print([i.duration_seconds for i in audio_chunks])

gold = "Recimo da je to prvi chunk. To drugi. In končno tretji."
model_output = "To je ta drugi."

from fuzzywuzzy import fuzz
def get_subset(gold, model_output):
    results = list()
    gold = gold.split(" ")
    model_output = model_output.split(" ")
    M = len(gold)
    for start in range(M):
        for end in range(M):
            try:
                subset = gold[start:end]
            except IndexError:
                continue
            ratio = fuzz.ratio(" ".join(model_output), " ".join(subset))
            results.append((ratio, start, end, " ".join(subset)))
    maximum_ratio = max(results, key=lambda tup: tup[0])
    for result in results:
        if result[0] == maximum_ratio[0]:
            return result
    return results


get_subset(gold, model_output)


get_subset(
    "a o tome govori i tematika zbog kojeg su se sve žalili",
    "aotomegovoritematika"
)

get_subset(
    "poštovane kolegice i kolege zastupnici molio bi da nastavimo sa radom sjednice",
    "poštovane kolegice ikolege zastupnici"
)

import difflib

l1 = "poštovane kolegice i kolege zastupnici molio bi da nastavimo sa radom sjednice"
l2 = "poštovane kolegice ikolege zastupnici"

seq_mat = difflib.SequenceMatcher(a=l1, b=l2)

match = seq_mat.find_longest_match(alo=0, ahi=len(l1), blo=0, bhi=len(l2))

print("============ Longest Matching Sequence ==================")
print("\nMatch Object : {}".format(match))
print("Matching Sequence from l1 : {}".format(l1[match.a:match.a+match.size]))
print("Matching Sequence from l2 : {}\n".format(l2[match.b:match.b+match.size]))

print("============ All Matching Sequences ==================")
for match in seq_mat.get_matching_blocks():
    print("\nMatch Object : {}".format(match))
    print("Matching Sequence from l1 : {}".format(l1[match.a:match.a+match.size]))
    print("Matching Sequence from l2 : {}".format(l2[match.b:match.b+match.size]))

from Bio import pairwise2
from Bio.Seq import Seq
seq1 = Seq("ACCGGT") 
seq2 = Seq("ACGT")

pairwise2.align.globalxx(seq1, seq2)

alignments = pairwise2.align.globalxx(
        Seq("a o tome govori i tematika zbog kojeg su se sve zalili".encode("ascii")),
        Seq("a o tome govori itematika zbog koje su se sve zalili")
)
alignments

alignments[1].seqB

l1 = "postovane kolegice i kolege zastupnici molio bi da nastavimo sa radom sjednice"
l2 = "postovane kolegice ikolege zastupnici ja bi molio"

alignments = pairwise2.align.globalxx(
        Seq(l2),
        Seq(l1))
alignments

alignments[1].seqB

get_subset(l1, l2)

get_subset(l1, "")

def Getletterfromindex(num):
    import string
    #produces a string from numbers so

    #1->a
    #2->b
    #26->z
    #27->aa
    #28->ab
    #52->az
    #53->ba
    #54->bb

    num2alphadict = dict(zip(range(1, 27), string.ascii_lowercase))
    outval = ""
    numloops = (num-1) //26

    if numloops > 0:
        outval = outval + self.Getletterfromindex(numloops)

    remainder = num % 26
    if remainder > 0:
        outval = outval + num2alphadict[remainder]
    else:
        outval = outval + "z"
    return outval

Getletterfromindex(26)

# Import pairwise2 module
from Bio import pairwise2

# Import format_alignment method
from Bio.pairwise2 import format_alignment

# Define two sequences to be aligned
X = "a o tome govori i tematika zbog kojeg su se sve zalili"
Y = "a o tome govori itematika zbog koje su se"

# Get a list of the global alignments between the two sequences ACGGGT and ACG
# No parameters. Identical characters have score of 1, else 0.
# No gap penalties.
alignments = pairwise2.align.globalxx(X, Y)

# Use format_alignment method to format the alignments in the list
for a in alignments:
    print(format_alignment(*a))
