==> 18_paper_related.ipynb <==
import pandas as pd
import os
kaldi = pd.read_csv("wav.scp", sep="\s+", names=["barename", "fullpath"])
kaldi["filename"] = kaldi.fullpath.apply(os.path.basename)
kaldi.head()

our = pd.read_csv("15_new_DS.csv")
our["filename"] = our.path.apply(os.path.basename)
our.head()

our["is_in_kaldi"] = our.filename.isin(kaldi.filename)

our_surplus = our_set.difference(kaldi_set)
kaldi_surplus = kaldi_set.difference(our_set)

print(f"We have {len(our_surplus)} that kaldi hasn't.")
print(f"Kaldi have {len(kaldi_surplus)} that we don't.")

len(our_surplus)/our.shape[0]

our_surplus
